DAOS-3564 dtx: cleanup DTX data before restart it
When hit MVCC conflict, current DTX will be required to
restart. For single RDG based DTX, it will be restarted
on leader with newer epoch. For distributed transaction,
the client will restarst the DTX. In spite of for which
case, we need to cleanup related DTX data (both the DTX
entry and the DTX handle) before the restart. Otherwise,
the subsequent inserting DTX entry may hit former stale
DTX information on the DTX participant(s) that have not
hit MVCC confict.

Signed-off-by: Fan Yong <fan.yong@intel.com>